Fixed NullPointerException, fixed Passenger.ssn mapping

+++ src/main/java/com/sedbproj/Railways/Services/UserService.java
@@ -12,7 +12,7 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
-    public Optional<UserEntity> getUserById(int id) {
+    public Optional<UserEntity> getUserById(Long id) {
         return userRepository.findById(id);
     }
 }
